Added support for targets
Tests can now be run on multiple targets in one call to plptest.
Targets are defined in testsets, for example pulp-open and
siracusa in pulp-sdk testsuites.
Targets to be included in the testset run are specified on command
line.
The testset run will duplicate the tests when several targets
are found.